audio: Refactor request handlers into a separate file
