+++ audio/firmware.c
@@ -1,19 +1,22 @@
-/*
- * Copyright (C) 2017 Kyle Robbertze
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+// Copyright (C) 2017 Kyle Robbertze <krobbertze@gmail.com>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+/** \file firmware.c
+ * Initialises the USB audio firmware and handles interrupts
  */
+
 #include <stdio.h>
 
 #include <fx2regs.h>
@@ -25,6 +28,7 @@
 #include <eputils.h>
 
 #include "fx2lights.h"
+#include "audiodata.h"
 
 #define SYNCDELAY SYNCDELAY4
 #define REARMVAL 0x80
@@ -112,11 +116,6 @@ void main() {
     }
 }
 
-// copied routines from setupdat.h
-BOOL handle_get_descriptor() {
-    return FALSE;
-}
-
 // value (low byte) = ep
 #define VC_EPSTAT 0xB1
 
@@ -137,40 +136,6 @@ BOOL handle_vendorcommand(BYTE cmd) {
         printf ( "Need to implement vendor command: %02x\n", cmd );
     }
     return FALSE;
-}
-
-// this firmware only supports 0,0
-BOOL handle_get_interface(BYTE ifc, BYTE* alt_ifc) { 
-    printf ( "Get Interface\n" );
-    if (ifc==0) {*alt_ifc=0; return TRUE;} else { return FALSE;}
-}
-BOOL handle_set_interface(BYTE ifc, BYTE alt_ifc) { 
-    printf ( "Set interface %d to alt: %d\n" , ifc, alt_ifc );
-
-    if (ifc==0&&alt_ifc==0) {
-        // SEE TRM 2.3.7
-        // reset toggles
-        RESETTOGGLE(0x02);
-        RESETTOGGLE(0x86);
-        // restore endpoints to default condition
-        RESETFIFO(0x02);
-        EP2BCL=0x80;
-        SYNCDELAY;
-        EP2BCL=0X80;
-        SYNCDELAY;
-        RESETFIFO(0x86);
-        return TRUE;
-    } else 
-        return FALSE;
-}
-
-// get/set configuration
-BYTE handle_get_configuration() {
-    return 1; 
-}
-
-BOOL handle_set_configuration(BYTE cfg) { 
-    return cfg==1 ? TRUE : FALSE; // we only handle cfg 1
 }
 
 // copied usb jt routines from usbjt.h
